Test contract & mock LINK dependencies
Adds some basic tests that utilise mocks to mock out the
Chainlink VRF. Also sets up test to environment to contain
the LINK contract.

Some aspects of the contract are yet to be tested.

The mocking of the Chainlink VRF and creation of the LINK
token contract is also done when spinning up local environment.
This is useful because in a local environment, we don't have access
to an actual Chainlink oracle, nor does the LINK token contract
come installed in our local environment.

To assist in achieving replicable deployments and easier to
write as well as speedier tests, switch to using the
`hardhat-deploy` plugin.