Manually send LINK instead of using `fund-link` task
The `fund-link` task produces logs that clutter up the
logs when we are testing. Removes the usage of `fund-link`
in the tests, and replaces this with a manual interaction
with the `LinkToken` contract.

Modify the `tsconfig.json` file so it inspects the types
present in `typechain-types`. This resolves an issue with
the ethers function `getContractAt`, where it was resolving
the return type as a generic `Contract` instead of a concrete
contract type.